Implement Point Light, update quicktest.py to include a Point Light
