+++ quicktest.py
@@ -2,7 +2,7 @@
 
 import math
 import os
-from term3d.core import term3d, Vec3
+from term3d.core import term3d, Vec3, PointLight
 from term3d.shpbuild import build_cube, build_uv_sphere, build_cylinder, build_torus
 from term3d.utils import set_mat
 try:
@@ -53,8 +53,16 @@ class OrbitScene:
         self.engine.set_render_quality(q)
 
         self.engine.set_clear_color(20,30,40) 
-        self.engine.add_light(Vec3(0.5, 0.7, -1.0), (255, 255, 255), 1.0)
-        self.engine.set_ambient_light(80,80,80) 
+        self.engine.add_light(Vec3(0.5, 0.7, -1.0), (255, 255, 255), 0.4)       # Make it low for now
+
+        # Add a point light, focus on the circle orbiting, if it become dark near siting between the cube and camera and lit(color toward orange) siting opposite the camera, the PointLight work!
+        point_light_pos = Vec3(0, 0, 0)   # I add a point light in center for testing
+        point_light_color = (255, 210, 150)
+        point_light_intensity = 1.0
+        new_point_light = PointLight(point_light_pos, point_light_color, point_light_intensity)
+        self.engine.lights.append(new_point_light)
+
+        self.engine.set_ambient_light(60,60,60)    
 
         self.orbit_angle_sphere = 0.0
         self.orbit_angle_cylinder = math.pi / 2
